Extract sync relay health helpers
